撮影日基準で並び替える
・撮影日はlong

+++ app/src/main/kotlin/net/mmho/photomap2/HashedPhoto.kt
@@ -5,15 +5,15 @@ import android.os.Parcelable
 
 import net.mmho.photomap2.geohash.GeoHash
 
-data class HashedPhoto(val photo_id:Long,val date_taken: Int,val hash:GeoHash) : Parcelable {
+data class HashedPhoto(val photo_id:Long,val date_taken: Long,val hash:GeoHash) : Parcelable {
     constructor(source: Parcel):
-        this(source.readLong(), source.readInt(), source.readParcelable<GeoHash>(GeoHash::class.java.classLoader))
+        this(source.readLong(), source.readLong(), source.readParcelable<GeoHash>(GeoHash::class.java.classLoader))
 
     override fun describeContents() = 0
 
     override fun writeToParcel(dest: Parcel?, flags: Int) {
         dest?.writeLong(photo_id)
-        dest?.writeInt(date_taken)
+        dest?.writeLong(date_taken)
         dest?.writeParcelable(hash, 0)
     }
